requirements.txt
# Core
python-dotenv>=1.0.0
pyyaml>=6.0.1
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
aiohttp>=3.9.0

# ML/CV
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
opencv-python-headless>=4.8.0
pillow>=10.0.0
numpy>=1.24.0

# Geospatial
shapely>=2.0.0
pyproj>=3.0.0  # Compatible with Python 3.8+

# Data Processing
pandas>=2.0.0
orjson>=3.9.0

# Logging & Monitoring
loguru>=0.7.0

# Security
cryptography>=41.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
